[Components] changes at multi components
