+++ app/components/shared/TxHistory/index.js
@@ -48,16 +48,17 @@ const TxHistory = ({
       const txTimestamp = tx.txTimestamp;
       // we define the transaction icon by its rowType, so we pass it as a
       // className props
-      let rowType = tx.status || tx.txType || tx.txDirection;
+      let rowType = tx.status || tx.txType;
       rowType = rowType.toLowerCase();
+      // If it is a regular tx we use its direction to show a proper icon.
+      if (rowType === txTypes.REGULAR) rowType = tx.txDirection;
       const Component = TxRowByType[rowType];
       if (Component === StakeTxRow && isRegular) return;
       if (Component === RegularTxRow && isStake) return;
 
       const txOutputAddresses =
-        tx.originalTx &&
-        tx.originalTx.outputs &&
-        tx.originalTx.outputs
+        tx.outputs &&
+        tx.outputs
           .filter((o) => !o.isChange)
           .map((o) => o.address)
           .join(" ");
@@ -72,7 +73,7 @@ const TxHistory = ({
             txTs: txTimestamp && tsDate(txTimestamp),
             overview,
             pending: tx.isPending,
-            onClick: () => history.push(`/transactions/history/${tx.txHash}`),
+            onClick: () => history.push(`/transaction/history/${tx.txHash}`),
             timeMessage: (txTimestamp) =>
               intl.formatMessage(timeMessageDefine.dayMonthHourDisplay, {
                 value: txTimestamp
