+++ app/components/views/TransactionsPage/HistoryTab/index.js
@@ -140,12 +140,14 @@ class History extends React.Component {
     ];
   }
 
-  getTransactions() {
-    return this.props.transactions;
-  }
-
   onLoadMoreTransactions() {
     this.props.getTransactions();
+  }
+
+  getTransactions() {
+    const { transactions } = this.props;
+    if (!transactions) return [];
+    return Object.keys(transactions).map((hash) => transactions[hash]);
   }
 
   onChangeFilter(value) {
